rename bangladesh to bgd
